+++ core/src/com/pacman/Pantallas/PantallaFinDelJuego.java
@@ -1,4 +1,4 @@
-package com.pacman;
+package com.pacman.Pantallas;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Color;
@@ -8,6 +8,7 @@ import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.ui.Skin;
 import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
+import com.pacman.JuegoPrincipal;
 
 public class PantallaFinDelJuego extends PantallaBase {
     //pantalla que indica que finalizo el juego y da la posibilidad de reiniciar la partida
